+++ src/app/modules/demo/demo.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-root',
+  selector: 'component-demo',
   templateUrl: './demo.component.html',
   styleUrls: ['./demo.component.scss'],
 })