+++ src/app/modules/hello-world/hello-world.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-hello-world',
+  selector: 'component-hello-world',
   templateUrl: './hello-world.component.html',
   styleUrls: ['./hello-world.component.scss'],
 })