+++ HISTORY.md
@@ -5,3 +5,4 @@
 1. Installed linters and what/not
 1. [`ng generate component "hello-world"`](https://angular.io/tutorial/first-app/first-app-lesson-01)
 1. Refactored `AppComponent` as `modules/demo/demo.component.ts`
+1. [`ng generate component Home`](https://angular.io/tutorial/first-app/first-app-lesson-02)